requirements.txt
aiogram>=3.4.0,<4.0.0
SQLAlchemy>=2.0.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.0.0,<9.0.0
aiohttp>=3.9.0,<4.0.0
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
pydantic>=2.6.0,<3.0.0
python-multipart
